+++ lifelines/__init__.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
 from .estimation import KaplanMeierFitter, NelsonAalenFitter, \
-    AalenAdditiveFitter, BreslowFlemingHarringtonFitter, CoxPHFitter
+    AalenAdditiveFitter, BreslowFlemingHarringtonFitter, CoxPHFitter, \
+    WeibullFitter, ExponentialFitter
 
 __all__ = ['KaplanMeierFitter', 'NelsonAalenFitter', 'AalenAdditiveFitter',
-           'BreslowFlemingHarringtonFitter', 'CoxPHFitter']
+           'BreslowFlemingHarringtonFitter', 'CoxPHFitter', 'WeibullFitter', 
+           'ExponentialFitter']